requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Environment Variables
python-dotenv==1.0.0

# MongoDB
motor==3.3.2
pymongo==4.6.0

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Face Recognition
face-recognition==1.3.0
dlib==19.24.2

# Object Detection (YOLO)
ultralytics==8.0.227

# Data Processing
numpy==1.24.3

# Pydantic for data validation
pydantic==2.5.0
